:bug: add arm64 support

+++ docker/mint/start.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
-export CODE="mint"
-export NAME="$CODE-$USER"
+set -e
+if [ "$#" -ne 1 ]; then
+    echo "USAGE: $0 PHP_VERSION"
+    exit 1
+fi
+
+export CODE="mint-$(uname -m)-$1"
 
 podman run --rm -it --events-backend=file --hostname=mint --network host -v $PWD:/workspace:z $CODE
